fixing memory leak

+++ playphrase/src/main/java/dragon/tamu/playphrase/FileAccessor.java
@@ -244,7 +244,7 @@ public class FileAccessor
         }
         phrase.addLanguage(language, newName.getPath());
 
-        saveInfoToFile(informationList);
+        saveInfoToFile(informationList, false);
 
         return phrase;
     }
@@ -273,7 +273,7 @@ public class FileAccessor
 
         if (phrase != null) {
             category.phraseList.remove(phrase);
-            saveInfoToFile(informationList);
+            saveInfoToFile(informationList, false);
         }
     }
 
@@ -296,7 +296,7 @@ public class FileAccessor
             // Adding a new category to the top will help keep Uncategorized at end
             // Also provides right functionality for the AddCategory button
             informationList.add(0, newCat);
-            saveInfoToFile(informationList);
+            saveInfoToFile(informationList, false);
         }
         return informationList;
     }
@@ -329,7 +329,7 @@ public class FileAccessor
             }
 
             informationList.remove(category);
-            saveInfoToFile(informationList);
+            saveInfoToFile(informationList, false);
         }
         return informationList;
     }
@@ -359,7 +359,7 @@ public class FileAccessor
         }
     }
 
-    public void saveInfoToFile(ArrayList<Category> categoryArrayList) {
+    public void saveInfoToFile(ArrayList<Category> categoryArrayList, boolean cleanFiles) {
         JSONArray categories = categoryToJSON(categoryArrayList);
 
         JSONObject saveJSON = new JSONObject();
@@ -370,35 +370,37 @@ public class FileAccessor
             e.printStackTrace();
         }
 
-        File directory = new File(context.getFilesDir().getAbsolutePath() + "/recordings");
-        if (directory.exists()) {
-            List<File> temp = Arrays.asList(directory.listFiles());
-            List<File> toDelete = new ArrayList<>(temp);
-            List<File> toKeep = new ArrayList<>();
+        if (cleanFiles) {
+            File directory = new File(context.getFilesDir().getAbsolutePath() + "/recordings");
+            if (directory.exists()) {
+                List<File> temp = Arrays.asList(directory.listFiles());
+                List<File> toDelete = new ArrayList<>(temp);
+                List<File> toKeep = new ArrayList<>();
 
-            for (Category cat : categoryArrayList) {
-                for (Object objPhrase : cat.phraseList) {
-                    Phrase phrase = (Phrase) objPhrase;
-                    for (String location : phrase.phraseLanguages.values()) {
-                        for (File check : toDelete) {
-                            if (check.getPath().equals(location)) {
-                                toKeep.add(check);
-                                break;
+                for (Category cat : categoryArrayList) {
+                    for (Object objPhrase : cat.phraseList) {
+                        Phrase phrase = (Phrase) objPhrase;
+                        for (String location : phrase.phraseLanguages.values()) {
+                            for (File check : toDelete) {
+                                if (check.getPath().equals(location)) {
+                                    toKeep.add(check);
+                                    break;
+                                }
                             }
                         }
                     }
                 }
-            }
-            for (File file : toKeep) {
-                toDelete.remove(file);
-            }
+                for (File file : toKeep) {
+                    toDelete.remove(file);
+                }
 
-            for (File file : toDelete) {
-                file.delete();
-            }
-            Log.d("saveInfoToFile", "delete check\n" + toDelete.toString() + "\n\n");
-            Log.d("saveInfoToFile", "delete check All files\n" + temp.toString() + "\n\n");
+                for (File file : toDelete) {
+                    file.delete();
+                }
+                Log.d("saveInfoToFile", "delete check\n" + toDelete.toString() + "\n\n");
+                Log.d("saveInfoToFile", "delete check All files\n" + temp.toString() + "\n\n");
 
+            }
         }
 
         try {
